requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.1
pluggy==1.2.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
pycparser==2.21
Pygments==2.16.1
pyroscope-io==0.8.5
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
rich==13.5.2
sentry-sdk==1.29.2
urllib3==2.0.4
Werkzeug==2.3.6
